@@ -27,15 +27,15 @@
     "demo:release": "npm run demo:build && node scripts/gh-pages"
   },
   "dependencies": {
-    "@angular/animations": "^15.2.9",
-    "@angular/cdk": "^15.2.9",
-    "@angular/common": "^15.2.9",
-    "@angular/compiler": "^15.2.9",
-    "@angular/core": "^15.2.9",
-    "@angular/forms": "^15.2.9",
-    "@angular/platform-browser": "^15.2.9",
-    "@angular/platform-browser-dynamic": "^15.2.9",
-    "@angular/router": "^15.2.9",
+    "@angular/animations": "^18.1.2",
+    "@angular/cdk": "^18.1.2",
+    "@angular/common": "^18.1.2",
+    "@angular/compiler": "^18.1.2",
+    "@angular/core": "^18.1.2",
+    "@angular/forms": "^18.1.2",
+    "@angular/platform-browser": "^18.1.2",
+    "@angular/platform-browser-dynamic": "^18.1.2",
+    "@angular/router": "^18.1.2",
     "cleave.js": "^1.6.0",
     "core-js": "^2.6.3",
     "date-fns": "^2.6.0",
@@ -43,31 +43,31 @@
     "rxjs": "~6.5.3",
     "semver": "^5.6.0",
     "tslib": "^2.5.3",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.14.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.2.8",
-    "@angular-eslint/builder": "15.2.1",
-    "@angular-eslint/eslint-plugin": "15.2.1",
-    "@angular-eslint/eslint-plugin-template": "15.2.1",
-    "@angular-eslint/schematics": "15.2.1",
-    "@angular-eslint/template-parser": "15.2.1",
-    "@angular/cli": "^15.2.8",
-    "@angular/compiler-cli": "^15.2.9",
-    "@angular/language-service": "^15.2.9",
+    "@angular-devkit/build-angular": "^18.1.2",
+    "@angular-eslint/builder": "18.1.0",
+    "@angular-eslint/eslint-plugin": "18.1.0",
+    "@angular-eslint/eslint-plugin-template": "18.1.0",
+    "@angular-eslint/schematics": "18.1.0",
+    "@angular-eslint/template-parser": "18.1.0",
+    "@angular/cli": "^18.1.2",
+    "@angular/compiler-cli": "^18.1.2",
+    "@angular/language-service": "^18.1.2",
     "@salesforce-ux/design-system": "~2.9.0",
     "@stackblitz/sdk": "^1.3.0",
     "@types/cleave.js": "^1.4.7",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "^2.0.6",
     "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "5.48.2",
-    "@typescript-eslint/parser": "5.48.2",
+    "@typescript-eslint/eslint-plugin": "^7.2.0",
+    "@typescript-eslint/parser": "^7.2.0",
     "app-root-path": "^2.2.0",
     "codelyzer": "^6.0.0",
     "conventional-changelog-cli": "^2.1.1",
     "del": "^3.0.0",
-    "eslint": "^8.33.0",
+    "eslint": "^8.57.0",
     "gh-pages": "^2.1.1",
     "gulp": "^4.0.2",
     "gulp-changed": "^3.2.0",
@@ -86,7 +86,7 @@
     "karma-jasmine-html-reporter": "^2.1.0",
     "karma-sauce-launcher": "^2.0.2",
     "markdown-it": "^8.4.2",
-    "ng-packagr": "^15.2.2",
+    "ng-packagr": "^18.1.0",
     "npm-run-all": "^4.1.5",
     "prismjs": "^1.23.0",
     "protractor": "~7.0.0",
@@ -97,7 +97,7 @@
     "shelljs": "^0.8.4",
     "simple-git": "^1.107.0",
     "ts-node": "~7.0.0",
-    "typescript": "~4.8.0"
+    "typescript": "~5.4.5"
   },
   "publishConfig": {
     "registry": "https://registry.npmjs.org/"
