@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule, } from '@angular/common';
 import { RouterModule, Routes } from '@angular/router';
 import { ReactiveFormsModule } from '@angular/forms';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { NglComboboxesModule, NglPillsModule } from 'ng-lightning';
 import { NglDemoExampleModule } from 'src/app/example/example.module';
 
@@ -17,21 +17,15 @@ const routes: Routes = [
   { path: '', component: DemoLookupsComponent },
 ];
 
-@NgModule({
-  imports: [
-    CommonModule,
-    RouterModule.forChild(routes),
-    ReactiveFormsModule,
-    HttpClientModule,
-    NglComboboxesModule,
-    NglPillsModule,
-    NglDemoExampleModule,
-  ],
-  declarations: [
-    DemoLookupsComponent,
-    DemoLookupsBasic,
-    DemoLookupsHttp,
-    DemoLookupsMultiple,
-  ],
-})
+@NgModule({ declarations: [
+        DemoLookupsComponent,
+        DemoLookupsBasic,
+        DemoLookupsHttp,
+        DemoLookupsMultiple,
+    ], imports: [CommonModule,
+        RouterModule.forChild(routes),
+        ReactiveFormsModule,
+        NglComboboxesModule,
+        NglPillsModule,
+        NglDemoExampleModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class NglDemoLookupsModule {}
